+++ CHANGELOG.md
@@ -24,6 +24,7 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 - fixed improves method setIptables from wirelessmode default mode 
 - fixed clear dependencies not used from requirements files
+- fixed include message: the module not found or failed to import.
 
 
 ## [released]